factor out function call graph computation
